trace_keyspace_helper::apply_events_mutation(): Avoid extra std::move of std::deque
events_records are promised to be kept alive till the future returned
by apply_events_mutation() resolves: it's dowithificated by a caller already.

In addition, since its passed by a reference, it's a logical thing to demand
it to be kept alive by a caller till the future above resolves.

Signed-off-by: Vlad Zolotarov <vladz@cloudius-systems.com>