database: make column family stats mutable
The make_reader method is currently a const method, but we would like to start
keeping hit statistics from it.

Instead of relaxing the const condition too much, we can just mark the _stats
field as mutable, indicating that make_reader will not be able to change
anything in the CF, except for keeping statistics.

Signed-off-by: Glauber Costa <glauber@scylladb.com>